Split sitemap a11y tests into pieces for future parallelization (#4484)
* Split sitemap a11y tests into pieces for future parallelization

* .
