+++ test/accessibility/sitemap-1.spec.js
@@ -1,0 +1,25 @@
+// nightwatch is capable of running tests in parallel, but only if the
+// test definitions are in separate files. This runs accessibility tests
+// over the first 1/4 of urls in the sitemap, while `sitemap-2-4.spec.js`
+// runs over the others. Crude, but this enables nightwatch to parallelize these.
+
+const Timeouts = require('../util/timeouts.js');
+const SitemapHelpers = require('../util/sitemap-helpers');
+
+module.exports = {
+  'sitemap 1/4': (client) => {
+    SitemapHelpers.sitemapURLs(urls => {
+      const mark = Math.ceil(urls.length / 4);
+      const segment = urls.splice(0, mark);
+
+      segment.forEach(url => {
+        client
+          .url(url)
+          .waitForElementVisible('body', Timeouts.normal)
+          .axeCheck('document', { scope: url });
+      });
+
+      client.end();
+    });
+  }
+};
